Extract `PerformerCreateInput` creation
