+++ ui/v2.5/src/components/Galleries/GalleryDetails/GalleryScrapeDialog.tsx
@@ -14,10 +14,9 @@ import {
   useStudioCreate,
   usePerformerCreate,
   useTagCreate,
-  stringToGender,
+  makePerformerCreateInput,
 } from "src/core/StashService";
 import { useToast } from "src/hooks";
-import { filterData } from "src/utils";
 
 function renderScrapedStudio(
   result: ScrapeResult<string>,
@@ -299,33 +298,7 @@ export const GalleryScrapeDialog: React.FC<IGalleryScrapeDialogProps> = (
   }
 
   async function createNewPerformer(toCreate: GQL.ScrapedScenePerformer) {
-    const input: GQL.PerformerCreateInput = {
-      name: toCreate.name,
-      url: toCreate.url,
-      gender: stringToGender(toCreate.gender),
-      birthdate: toCreate.birthdate,
-      ethnicity: toCreate.ethnicity,
-      country: toCreate.country,
-      eye_color: toCreate.eye_color,
-      height: toCreate.height,
-      measurements: toCreate.measurements,
-      fake_tits: toCreate.fake_tits,
-      career_length: toCreate.career_length,
-      tattoos: toCreate.tattoos,
-      piercings: toCreate.piercings,
-      aliases: toCreate.aliases,
-      twitter: toCreate.twitter,
-      instagram: toCreate.instagram,
-      tag_ids: filterData((toCreate.tags ?? []).map((t) => t.stored_id)),
-      image:
-        (toCreate.images ?? []).length > 0
-          ? (toCreate.images ?? [])[0]
-          : undefined,
-      details: toCreate.details,
-      death_date: toCreate.death_date,
-      hair_color: toCreate.hair_color,
-      weight: toCreate.weight ? Number(toCreate.weight) : undefined,
-    };
+    const input = makePerformerCreateInput(toCreate);
 
     try {
       const result = await createPerformer({
